refactor: decouple variables some more for treeshaking
